+++ rampwf/prediction_types/__init__.py
@@ -4,14 +4,12 @@ from .detection import make_detection
 from .multiclass import make_multiclass
 from .regression import make_regression
 from .generative_regression import make_generative_regression
-from .generative_regression_dists import make_generative_regression_dists
 
 __all__ = [
     'make_clustering',
     'make_combined',
     'make_detection',
+    'make_generative_regression',
     'make_multiclass',
     'make_regression',
-    'make_generative_regression',
-    'make_generative_regression_dists'
 ]
