+++ rampwf/workflows/__init__.py
@@ -7,13 +7,12 @@ from .feature_extractor import FeatureExtractor
 from .feature_extractor_classifier import FeatureExtractorClassifier
 from .feature_extractor_regressor import FeatureExtractorRegressor
 from .image_classifier import ImageClassifier
+from .grid_feature_extractor import GridFeatureExtractor
+from .generative_regressor import GenerativeRegressor
 from .simplified_image_classifier import SimplifiedImageClassifier
 from .object_detector import ObjectDetector
 from .regressor import Regressor
 from .ts_feature_extractor import TimeSeriesFeatureExtractor
-from .grid_feature_extractor_classifier import GridFeatureExtractorClassifier
-from .generative_regressor_selfcorellated import GenerativeRegressorSelf
-from .generative_regressor_selfcorellated_d import GenerativeRegressorSelfDist
 
 __all__ = [
     'AirPassengers',
@@ -24,12 +23,11 @@ __all__ = [
     'FeatureExtractor',
     'FeatureExtractorClassifier',
     'FeatureExtractorRegressor',
+    'GenerativeRegressor',
+    'GridFeatureExtractorClassifier',
     'ImageClassifier',
     'SimplifiedImageClassifier',
     'ObjectDetector',
     'Regressor',
     'TimeSeriesFeatureExtractor',
-    'GridFeatureExtractorClassifier',
-    'GenerativeRegressorSelf',
-    'GenerativeRegressorSelfDist'
 ]
